+++ src/JuMPChance.jl
@@ -132,7 +132,7 @@ end
 type ChanceConstr
     ccexpr::CCAffExpr
     sense::Symbol # :(<=) or :(>=), right-hand side assumed to be zero
-    with_probability::Float64 # with this probability *or less*
+    with_probability::Float64 # with this probability *or greater*
     uncertainty_budget_mean::Int # for now, with Bertsimas-Sim uncertainty
     uncertainty_budget_variance::Int # for now, with Bertsimas-Sim uncertainty
 end
@@ -142,6 +142,11 @@ ChanceConstr(ccexpr::CCAffExpr,sense::Symbol) = ChanceConstr(ccexpr, sense, NaN,
 function JuMP.addConstraint(m::Model, constr::ChanceConstr; with_probability::Float64=NaN, uncertainty_budget_mean::Int=0, uncertainty_budget_variance::Int=0)
     if !(0 < with_probability < 1)
         error("Must specify with_probability between 0 and 1")
+    end
+    if with_probability < 0.5
+        Base.warn_once("The meaning of with_probability has flipped! Constraints should be satisfied with given probability or greater. For backwards compatibility, this constraint will be interpreted in the old sense.")
+        with_probability = 1-with_probability
+        constr.sense = (constr.sense == :(>=)) ? :(<=) : :(>=)
     end
     constr.with_probability = with_probability
     constr.uncertainty_budget_mean = uncertainty_budget_mean